// src/app/shared/models/customer-apiresponse.ts
export interface Customer {
  Category: string;
  UpdatedDate: Date;
  Suppl_RegistrationNumber: string;
  AuthorisedSignatory: string;
  purpose: string;
  Rating: string;
  OperatingUnit: string;
  Internal: string;
  Legacy_ID: null;
  RelatedPartyRef: string;
  SubIndustry: string;
  OTASalesPerson: string;
  StartDate: Date;
  UpdatedBy: string;
  DealCategory: string;
  CustomerSubCategory: string;
  EndDate: Date;
  Freeze: string;
  SiteUseCode: string;
  CUSTOMERID: string;
  RelationshipWithTA: string;
  Status: string;
  Industry: string;
  StatusWF: string;
  CustomerTaxReference: string;
  PrimarySiteUseFlag: string;
  FreezeDescription: null;
  Cust_RegistrationNumber: string;
  IsSupplier: string;
  ServicesToBeProvided: string;
  CreatedBy: null;
  AccountRelationship: string;
  CustomerCompany: null;
  ShortName: string;
  BusinessName: string;
  IsCustomer: string;
  Code: null;
  Type: string;
  CreatedDate: null;
  Markup: string;
  CollectionAgent: string;
  PersonFlag: string;
}